Add support for Marketplace
This commit adds support to marketplace module. In order to support the file upload to the files API a few architectural changes were made to reuse the existing `FilesClient` and avoid code replication. The redundant `transport` interface/implementation classes were deleted because they have shown to be redundant in the actual implementation. The layer API Client was refactored to handle different API endpoints that are set as a `baseUri` before instantiation.

Out of the scope, `PaymentSenderType` was created in the Payments module, and Instruments was refactored to user `AccountHolderType` enum from commons. A few more changes were performed to increase the code quality.